Separate pull request page in simpler components
